# Conflicts:
#	src/main/kotlin/com/lykke/matching/engine/messages/MessageProcessor.kt
#	src/main/kotlin/com/lykke/matching/engine/order/GenericLimitOrderProcessor.kt
#	src/main/kotlin/com/lykke/matching/engine/order/StopLimitOrderProcessor.kt
#	src/main/kotlin/com/lykke/matching/engine/services/AbstractLimitOrdersCancelService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/BalanceUpdateService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/CashInOutOperationService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/CashOperationService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/MarketOrderService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/MultiLimitOrderService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/ReservedBalanceUpdateService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/ReservedCashInOutOperationService.kt

+++ src/main/kotlin/com/lykke/matching/engine/fee/DefaultFeeCoefCalculator.kt
@@ -1,5 +1,7 @@
 package com.lykke.matching.engine.fee
 
+import java.math.BigDecimal
+
 class DefaultFeeCoefCalculator : FeeCoefCalculator {
-    override fun calculate(): Double? = null
+    override fun calculate(): BigDecimal? = null
 }
