+++ src/main/kotlin/com/lykke/matching/engine/services/BalanceUpdateService.kt
@@ -15,6 +15,7 @@ import com.lykke.utils.logging.MetricsLogger
 import org.apache.log4j.Logger
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 import java.util.Date
 
 @Service
@@ -40,12 +41,12 @@ class BalanceUpdateService @Autowired constructor (private val balancesHolder: B
             val reservedBalance = balancesHolder.getReservedBalance(message.clientId, message.assetId)
 
             val updated = balancesHolder.updateBalance(messageWrapper.processedMessage(),
-                    message.clientId, message.assetId, message.amount)
+                    message.clientId, message.assetId, BigDecimal.valueOf(message.amount))
             messageWrapper.processedMessagePersisted = true
             if (updated) {
                 balancesHolder.sendBalanceUpdate(BalanceUpdate(message.uid.toString(),
                         MessageType.BALANCE_UPDATE.name, Date(),
-                        listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, message.amount, reservedBalance, reservedBalance)), messageWrapper.messageId!!))
+                        listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, BigDecimal.valueOf(message.amount), reservedBalance, reservedBalance)), messageWrapper.messageId!!))
             }
             messageWrapper.writeResponse(ProtocolMessages.Response.newBuilder())
 
@@ -66,7 +67,7 @@ class BalanceUpdateService @Autowired constructor (private val balancesHolder: B
             }
 
             val updated = balancesHolder.updateBalance(messageWrapper.processedMessage(),
-                    message.clientId, message.assetId, message.amount)
+                    message.clientId, message.assetId, BigDecimal.valueOf(message.amount))
             messageWrapper.processedMessagePersisted = true
             if (!updated) {
                 messageWrapper.writeNewResponse(
@@ -79,7 +80,7 @@ class BalanceUpdateService @Autowired constructor (private val balancesHolder: B
             balancesHolder.sendBalanceUpdate(BalanceUpdate(message.uid,
                     MessageType.BALANCE_UPDATE.name,
                     Date(),
-                    listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, message.amount, reservedBalance, reservedBalance)),
+                    listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, BigDecimal.valueOf(message.amount), reservedBalance, reservedBalance)),
                     messageWrapper.messageId!!))
 
             messageWrapper.writeNewResponse(ProtocolMessages.NewResponse.newBuilder().setStatus(MessageStatus.OK.type))
