+++ src/test/kotlin/com/lykke/matching/engine/performance/MarketOrderPerformanceTest.kt
@@ -6,6 +6,7 @@ import com.lykke.matching.engine.utils.MessageBuilder
 import com.lykke.matching.engine.utils.PrintUtils
 import org.junit.Ignore
 import org.junit.Test
+import java.math.BigDecimal
 
 @Ignore
 class MarketOrderPerformanceTest: AbstractPerformanceTest() {
@@ -109,7 +110,8 @@ class MarketOrderPerformanceTest: AbstractPerformanceTest() {
 
         testBackOfficeDatabaseAccessor.addAsset(Asset("USD", 2))
         testBackOfficeDatabaseAccessor.addAsset(Asset("EUR", 2))
-        testDictionariesDatabaseAccessor.addAssetPair(AssetPair("EURUSD", "EUR", "USD", 5, 0.1, 0.2))
+        testDictionariesDatabaseAccessor.addAssetPair(AssetPair("EURUSD", "EUR", "USD",
+                5, BigDecimal.valueOf(0.1), BigDecimal.valueOf(0.2)))
         initServices()
 
         counter.executeAction { marketOrderService.processMessage(MessageBuilder.buildMarketOrderWrapper(MessageBuilder.buildMarketOrder(volume = 0.09))) }
